requirements.txt
## Requirements for spotify_images
pillow
spotipy
opencv-contrib-python